Add files via upload
Adjusted all the files to get the model to actually function.
With the current default values the population decreases to zero relatively quickly, which likely means the defaults are bad. But we should also check whether the code is working as intended. Of note, the herbivore population does NOT decline to zero alongside the tree population, which means my adjustment to the grazing submodel wasn't quite right in some way.